Honor token visibility in overview polling
